feat: add `valueOf` method to Graph, Line, and PrimaryAxis
Use `valueOf()` when you want to know the y-axis value for a specific data point, indicated by its zero-based index within the dataset.

The `valueOf()` method of Graph also accepts a 2nd argument which specifies the "view" to get the value from. For example, if two lines exist on the graph, you must pass `1` as the 2nd argument if you want the value of the 2nd line. To get the value of the 1st line, you can pass `0` or nothing.

If the given index is out of bounds or no data exists, the returned value is `null`.