feat: add `fill` method to Line
This method uses the new `UnderFill` class to fill the space under the line with a solid color or gradient. For a solid color, pass a string to `fill()`. For a gradient, pass an array of objects or 2+ object arguments. The gradient objects may contain a `color` string, an `opacity` float, an `offset` float, and/or an x/y pair of numbers.

- The first gradient object has a default `offset` of 0 and a default x/y of `x: 0, y: 0`
- The last gradient object has a default `offset` of 1
- Always specify an `offset` for gradient objects in between the first and last
- Specifying an x/y pair for gradient objects other than first/last is not supported

NOTE: In the future, an `InnerFill` class will be added.