rework view adding
Remove the Graph class's knowledge of an `el` property on each view. Instead, call the `_attach` method of each view, passing in the graph's element. Along with mounting itself to the graph, the view can perform additional actions in response to being attached. The `_graph` property will be set before the view's attach method is called.